Support randomized simulators, with Maze in view
This includes updating all other example models to return constant
distros. I also took a chance to make small refactorings in the examples
I needed to touch.

Also distraction has been renamed to epsilon (to follow Sutton and Barto,
and to avoid confusion with the distraction in the Maze example).