concrete.examples.mountaincar: An attempt to fix random initialization
There is a problem that applying repeat to singleton in tests does not
work (a confusing aspect of call-by-name semantics and side effects).
This requires a more serious fix, but right now, I am just trying to
find a pragmatic solution so that the reinformcement learning proceeds.
Randomized should be fixed for real later.