examples.concrete.simplemaze: Decrease flakiness of the experiment
There is an issue with the policy at positions in column 3 which is
difficult to learn correctly (it is quite unclear what policy is
optimal). I played with rewards a bit, and also made the test more
forgiving. It seems that now it fails on average once in 7-8 runs. Which
is not great but better than before.